+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Bump version of MariaDB; recommended version is `v10.5.x`
+-   Upgrade base container image Debian GNU/Linux from 10 `buster` to 11 `bullseye`
 
 ### Removed